+++ packages/test-factory/CHANGELOG.md
@@ -1,5 +1,14 @@
 # orchid-orm-test-factory
 
+## 0.1.9
+
+### Patch Changes
+
+- Updated dependencies
+  - pqb@0.7.6
+  - orchid-orm@1.3.9
+  - orchid-orm-schema-to-zod@0.1.6
+
 ## 0.1.8
 
 ### Patch Changes